backend/pyproject.toml
[project]
name = "garbanzo-ai-backend"
version = "0.1.0"
description = "FastAPI backend for Garbanzo AI Flutter app"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
